myproject/requirements.txt
asgiref==3.8.1
astroid==2.15.5
attrs==21.2.0
Automat==20.2.0
autopep8==2.3.1
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
botocore==1.34.113
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.1.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
dill==0.3.8
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-environ==0.11.2
djangorestframework==3.15.2
ec2-hibinit-agent==1.0.0
exceptiongroup==1.2.2
filelock==3.16.1
flake8==7.1.1
git-remote-codecommit==1.17
gunicorn==23.0.0
hibagent==1.0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
ikp3db==1.4.2
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
isort==5.12.0
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
MarkupSafe==2.0.1
mccabe==0.7.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
parso==0.8.4
pexpect==4.8.0
platformdirs==4.2.2
pluggy==1.5.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycodestyle==2.12.1
pyflakes==3.2.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.17.4
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.8.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.3.3
pytest-django==4.9.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
sqlparse==0.5.0
ssh-import-id==5.11
systemd-python==234
tomli==2.0.1
tomlkit==0.12.5
Twisted==22.1.0
typing_extensions==4.12.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.20
virtualenv==20.26.5
wadllib==1.3.6
whitenoise==6.7.0
wrapt==1.16.0
zipp==1.0.0
zope.interface==5.4.0
